Chatbox conditional redering done, socket connect isnt working
